Allow automatic flashing
If FLASH_ON_START variable is defined, the service will attempt to flash the specified
firmware on start but only if this firmware is not yet reported to be present on the co-processor.

Also, refactored the code in the index.js, extracting the flashing logic into a separatee module
and cleaning up sleep command promises.

Change-type: minor
Signed-off-by: Roman Mazur <roman@balena.io>